+++ CHANGELOG.md
@@ -475,6 +475,7 @@
 ## [unreleased]
 
 - On the user administration page, BEIS now appears as a separate organisation to avoid users being assigned to this org by accident
+- Users can upload activities in bulk from a CSV
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-29...HEAD
 [release-29]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-28...release-29